requirements.txt
asgiref
certifi==2020.12.5
cffi==1.14.4
channels
chardet==4.0.0
daphne
defusedxml==0.6.0
Django==2.1.5
django-adminlte3==0.1.6
django-allauth==0.44.0
django-bootstrap==0.2.4
django-ckeditor==6.0.0
django-crispy==1.8.1
django-js-asset==1.2.2
django-markdown-deux==1.0.5
fontawesome==5.10.1.post1
idna==2.10
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.7.7
pycparser==2.20
PyJWT==2.0.1
PyMySQL==1.0.2
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
